requirements.txt
cassandra-driver
pytest
tblib

